Added vue-router Fixed components to match new routes
