+++ src/components/views/About.vue
@@ -1,0 +1,3 @@
+<template>
+  <h1>About</h1>
+</template>
