+++ src/components/views/Home.vue
@@ -1,0 +1,7 @@
+<template>
+  <h1>Home page</h1>
+  <router-link to="/about">About</router-link>
+  <br />
+  <router-link to="/list">List</router-link>
+  <router-view />
+</template>
