+++ src/main.js
@@ -1,4 +1,7 @@
-import { createApp } from 'vue'
-import App from './App.vue'
+import { createApp } from "vue";
+import App from "./App.vue";
+import router from "@/router";
 
-createApp(App).mount('#app')
+createApp(App)
+  .use(router)
+  .mount("#app");
